Downcast ByteBuffer to Buffer
This CL solves the java error message, 'java.lang.NoSuchMethodError: No virtual
method position(I)Ljava/nio/ByteBuffer; in class Ljava/nio/ByteBuffer;'.

In Java 8 used by Wolvic, the class ByteBuffer inherits the position method
from the class Buffer and has a return type of java.nio.Buffer. On Java 11 and
lather used by chromium, the method is overridden with an implementation that
returns java.nio.ByteBuffer.

So, this changes to do explicitly the downcast from the class ByteBuffer to the
class Buffer for the backward compatibility according to the Java version.