Returns browser context based on content::BrowserContext
In WolvicContentBrowserClient, this changes the return type of the
single browser context for Wolvic.

Instead of returning WolvicBrowserContext, now it will return
content::BrowserContext. This more coherent to other content
implementations within Chromium tree (e.g. App Shell) and will be handy
when initializing content::ShellDevToolsManagerDelegate in the following
commit.