Implement User Agent settings (#19)
Add a mechanism of sending Wolvic settings to the Chromium backend and
also implements support for the User Agent settings. Logic for the User
Agent support is using the same approach as the gecko backend
(SessionImpl.java).

Additionally, remove GetUserAgentBasedOnPolicy() method since we're not
planning to support any enterprise policies and our implementation of
GetUserAgent() will be used by default.