Add SessionSettings::GetDefaultUserAgent()
Added a new method that retrieves the UA to be used by Chromium for
a given mode (mobile, desktop or VR). Part of the existing code
has to be moved from the content browser client to session settings
in order to allow both Wolvic and the WolvicContentBrowserClient
to access it.