wolvic: Change method names and add comments
This changes some method names, adds comments and a few logging
information to WvrManager implementation.

There are no funcional changes in this commit. It only serves to pave
the way mostly to resemble our implementation to the GVR and ARCore ones
(which we'll be referring our implementation a lot on them).