wolvic: Break down processing frame code in separated methods
This breaks down the processing frame state in a new method,
DrawFrameSubmitNow.

This also makes sure (by DCHECKing) that when OnWebXrFrameAvailable
runs, the frame arriving is still in the processing state and
guaranteeing the correct order functioning as defined by
WebXrPresentationState. Besides, it removes an unused member of the main
class (last_frame_index_).