wolvic: Add GpuFence for frame sync
GpuFence is a wrapper for native fence sync including IPC and mojo
transport.

render_completion_fence is used to avoid overstuffed buffers in WebXR.
The basic idea is to submit incomplete frames to Wolvic, with the fence
ensuring the next frame's GL commands get in line behind the current
frame.

Also, for tracing purposes, and if we want to do it later, it's simple
now to get the time when rendering completed for it.

GpuFence is needed for shared buffer transport on WebXR, coming next.