Revamp input source management
The input source management was not correct. There were some bugs
in the current implementation and there are also some missing very
important features.

This PR implements the following:
* Use the XrStandardGamepadBuilder to create the Gamepad object.
This ensures that is created with the xr-standard mapping required
by the WebXR implementation. Also it simplifies a bit the code.
* Properly fill in the primary_input_* and primary_squeeze_*
data. Those 4 values are defined in the WebXR spec (primary action
and primary squeeze action) and are used to properly trigger
the select* and squeeze* events. This requires storing the
previous state of the controllers.
* Fix touchpad and thumbstick support. There was no touchpad
support at all. Also axes values were not properly filled in.
Last but not least, thumbstick's click/touch/value were not
filled in.

Note: the index for where the data for a particular button is
located follows the xr-standard gamepad mapping that can be found
https://www.w3.org/TR/webxr-gamepads-module-1/#xr-standard-gamepad-mapping