Workaround to support older versions of three.js
Older versions of three.js (until r125) have the following issue
https://github.com/mrdoob/three.js/issues/21126 present which breaks
certain WebXR experiences like "moonrider" and "construct chess".

This workaround makes sure that any WebGL context is marked as XR
compatible when possible which means that applications no longer have
to mark it manually and in the correct order. This fix is verified to
work on both "moonrider" and "construct chess" and also doesn't break
sample WebGL pages that don't use WebXR.