Add support for WebXR Hand Input Module - Level 1
https://www.w3.org/TR/webxr-hand-input-1/

Since the support is already present in blink, we just need to fetch
the hand-tracking data pushed to as part of system state updates, and
populate it in the corresponding mojom data structures.

We also need to advertise the `generic-hand-select` profile in the
input source description, as required by the spec.