Implement DRM support
This change implements DRM support (for widevine and clearkey DRMs)
using the same logic as the android_webview. We don't download or store
any proprietary blobs, instead we use the Android DRM framework.

We also automatically grant the PROTECTED_MEDIA_IDENTIFIER permission
here, which should be changed to use the permission request dialog in
the future.