Add VisitedLinkReader and VisitedLinkWriter
These two components are used to exchange information about visited
links between the browser and the renderer. The writer is the one
that receives the list of visited URLs and stores them (actually
a hash of the URL) in a hash table for rapid searches. The reader
on the other hand, will check whether a given link found in a
web page was ever visited to show it later in a different colour.

Adding them is not enough to get the visited links functionality
working though. We still need 2 components:
* a WebContents observer tracking the visited links
* a class that exposes the VisitedLinkReader interface to the browser