Adding WolvicContents to track visited URLs
This new class is basically a WebContentsObserver used to capture
the list of visited URLs and pass them to WolvicBrowserContext so
that they could be stored as visited. This object is owned by an
utility class called WolvicContentsUserData that is stored as
user data inside the WebContents object so that we could easily
access it. As it is not passed to Java the ownership is kept
in that user data class.