Store WebContents in WolvicContents
The current code does not keep a reference to WebContents on the
native side. It's using the same structure as WebContentsFactory's
CreateWebContents(), but the difference is that we don't end
up calling TabAndroid::InitWebContents, and thus the WebContents
instance is not properly stored in a unique_ptr.

Use the recently added WolvicContents object to keep track of it.