Store the WebContentsDelegate in WolvicContents
The current code keeps the delegate instance in an static unique_ptr
in tab_jni.cc. That's unfortunate as it will be shared by all the
WebContent instances and might be the source of weird behaviours.

Instead it's better to use the recently added WolvicContents
object to keep the unique reference to the delegate.