Notify clients when the web engine wants to create a new Window
Certain navigations (like target=_blank links) require the web
engine to create a new Window. In those cases Chromium creates a
new WebContents object and notifies the client (Wolvic in this case)
about that. However the current architecture of Wolvic does
not allow us to properly use that new WebContents object to
create a new tab.

To circumvent that, we can use the fact that Chromium cancels
the creation of a new window if the ownership of the WebContents
object is not taken when AddNewContents() is called on the delegate.
We precisely do that (we let the new WebContents die) but we
use the target_url (the URL of the new window) to notify the
client (Wolvic) via a new method called onCreateNewWindow() that
was added to a new object called WolvicWebContentsDelegate which
directly inherits from WebContentsDelegateAndroid (the one we
were using previously).

The client just needs to override that onCreateWindow() method
and trigger the creation of a new window there using the passed
in URL.