Use generic-hand profile instead of generic-hand-select
The former does have a 3D model in the WebXR repository while the
latter don't, that's why we don't show the hand model correctly
in some experiences.

Apart from that do insert the hand profile at the beginning of the
list of profiles. Otherwise the physical controller profile would
be selected first.