Replace WebVr references by WebXR
A legacy rationale is behind the usage of WebVr in the manager APIs.
Some other APIs in Chromium were using the name of the old VR standard
for the Web, but there is really no reason to use that instead of
WebXR which is the one that arrived to stay.