Use gfx::Size instead of separate integers for WvrAPI
Instead of passing two integer arguments we can use a gfx::Size
which already encapsulates an integer size in a single attribute.
On the one hand it's more meaningful as it is now obvious that
both sizes are related, and on the other hand it allows us
to use the gfx::Size which is returned directly by other APIs
without having to decompose it in its two components.