Use an enumerated type in PushState
It used to receive a boolean as single argument. The problem of
using booleans in public APIs is that the caller does not really
know what it does really mean, forcing developers to check the
signature an implementation of the API. By using an enumerated
type the callsites become self-explanatory.