Replace a DCHECK by a null check
The call to GetJavaDelegate could return jobject whose obj() is null.
That's why all the delegates in chromium code have a is_null()
check after the aforementioned call. We instead had a DCHECK
because we thought it was not possible to get a null object there
but the truth is that sometimes we do and in those cases
it was causing a crash on release builds.

It's still unclear why it happens but we should avoid those
crashes. When that obj() is null then we bail out and the
navigation does not succeed. We haven't figured out yet which
cases trigger those conditions too.