Fix WebXR's spinning issue
This patch changes to use only the eye height for the space's
standing height from the floor so that fixes the spinning issue when
loading Archery Evolution WebXR game from the heyvr.io site.

Also, the standing height is multiplied by -1 according to the
patch reflection of crrev.com/c/2278317.