Add PicoNeo3 support
Wolvic already had support for that device. The problem was not that
the Neo3 was not properly select but also all the other device
whose enumerated type definitions were bellow the PicoNeo3 like
the Quest3 or the MagicLeap2 that were not selecting the proper
WebXR profile and thus loading the incorrect 3D model.