Do not reset the reference space on every frame
Resetting the reference space must be done only when the origin
changes signigicantly, for example as a response to a long click
on the menu button of a controller (like when recentering the view).

The code was doing it on every frame unconditionally instead. This
is wrong and was causing weird behaviours in several WebXR
experiences.

Fixes Wolvic's https://github.com/Igalia/wolvic/issues/1296