Enable AR support in the Wolvic device
This involves a few different changes. So far we have been
forcing the blend mode to opaque because we were only
supporting immersive-vr WebXR sessions. In order to
properly support AR cases we need to select the blend mode
depending on the session type and also the available blend
modes provided by Wolvic. Also the device should advertise
itself as AR capable.

However that is not enough for devices like Meta's Quest
because they only support opaque blend modes although
passthrough is supported via a special compositor layer.
For those devices we need to notify Wolvic that the browser
wants to start an immersive AR session so Wolvic could
decide the exact mechanism to use, either an AR compatible
blend mode (like for Pico4E) or a passthrough compositor
layer (like in the case of Quest3).