Implement a WebDataService factory
The WebDataServiceWrapper is a KeyedService that owns multiple
WebDataServices. One of those is the GetPaymentManifestData
which implements the logic to retrieve the WebApp's Manifests,
used by the WebPayment APIs when trying to find and launch
payment apps.

The WebDataServiceWrapperFactory is an abstract class that needs
to be implemented by Wolvic. This change defines a new class
WebDataServiceWrapperFactory precisely for that purpose. This
provides Wolvic with the specialized KeyedService factory to
create instances of the WebDataServices mentioned before.