Add WolvicBrowserContext java class
This patch adds WolvicBrowserContext java class that is an
wrapper that allows passing WolvicBrowserContext reference around
in the Java layer. It has an interface to get BrowserContextHandle
interface that provides a pointer to a native BrowserContext, and
can be used to by components to take a BrowserContext reference
within Java code.