Select the correct blend mode for WebXR sessions
The PickEnvironmentBlendModeForSession() call used to select the best
blend mode for the device depending on the platform supported blend modes.
However that was not totally matching the spec[1].

The blend mode required by the WebXR session is not the same as the
blend mode that the app will use. For example, Meta devices do only
support the opaque blend mode. That's why that method was selecting
opaque even for the "immersive-ar" session mode. However the spec
(and thus Chromium) was expecting for that device (a pass-through
device by the spec definitions) to select the alpha blend mode
for the WebXR session (even if the device would use opaque anyway
because that's the only one supported by the OpenXR runtime).

This is why in Meta devices the WebXR content for AR sessions was
actually opaque. Wolvic was properly initializing the passthrough
layer but the WebXR content (which is shown on top) was fully
opaque.

As part of the fix, we're selecting the additive blend mode for
devices supporting it (like MagicLeap2) regardless of the session
as per specs.

[1] https://www.w3.org/TR/webxr-ar-module-1/#xr-compositor-behaviors