Set the default user agent to the network context
This patch sets the desktop user agent by the default to the network
context so that we send the network request with the valid user
agent in the http header instead of empty string.

Since we have been sending empty UA values ​​so far, sending desktop
UA will get a better response from servers. Also, since Meta browser
always uses only desktop UA, the impact of this value will be low.

This fixes the payment request issue by empty user agent string
when accessing the google pay server. This approach will be simplest
and also we are considering other approaches.
- Set the desktop UA to only payment's the network request.
- Override UA before going out the network reqeust on demend.

We will revisit this to set the setting value to the network
context's user agent.