[WebPayments] WolvicPaymentUiService implements the PaymentRequestUI.Client interface
Now that the WolvicPaymentUiService implements the PaymentRequestUI.Client interface
we can use it to control the PaymentHandler window's life-cycle. In order to do this,
we need to create the PaymentRequestUI instance in Chromium, and getting it in Wolvic
from the WebContents received in the onCreateNewPaymentHandler.

The current design didn't allow to call the WebContentImpl::getOrSetUSerData method,
becuase it'd require the PaymentRequestUI to have a sinple consrtuctor with just one
WebContents parameter. Additionally, this WebContents instance might be already
initialized, which is not the case of our current design.

So the solution to implement the PaymentRequestUI::getFromContents was to add a new
public method in the WebContentsImpl class to directly set userData as an already
created object.