Rebase M131: Fixed crash in WolvicPermissionManager
In the CompleteRequest method we are removing the in_progress_request
from the vector own by the WolvicPermissionManager instance. An
iterator is obtained by looking into the vector, based on the pointer
obtainer from the Java side and we use this iterator to remove the
element once the request has been processed completely.

The problem is that as result of the "callback" execution, a new
element is added to the InProgressRequest vector, which causes the
iterator we obtained before to become invalid.

This change moves the logic to obtain the iterator after the callback
is executed and the vector is not modified anymore.